pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chemeleon2"
version = "0.1.0"
description = "A generative machine learning framework for crystal structure prediction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ase>=3.26.0",
    "average-minimum-distance>=1.6.0",
    "fire>=0.7.1",
    "h5py>=3.14.0",
    "hydra-colorlog==1.2.0",
    "hydra-core==1.3.2",
    "hydra-optuna-sweeper==1.2.0",
    "jupyterlab>=4.4.6",
    "lightning>=2.5.4",
    "peft>=0.17.1",
    "pydantic>=2.11.7",
    "pymatgen>=2025.6.14",
    "rich>=14.1.0",
    "timm>=1.0.19",
    "torch>=2.6.0",
    "torch-ema>=0.3",
    "torch-geometric>=2.6.1",
    "torchmetrics>=1.8.1",
    "tqdm>=4.67.1",
    "wandb>=0.21.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"] 
